Make post-wal commands conditional
This commit makes the post-wal script a bit more aware of its
environment. This is only to support Plasma with bspwm, which
needs the bspwm settings but not the dunst settings.

Since bspwm with KDE settings performs exceptionally well, I may drop
Plasma with bspwm support altogether and revert this commit at a later
date. The overhead of wmctrl and grep may not be worth it.

Additionally, the filename should probably be changed to a more
generic one in order to support more software as needed.

+++ .config/wal/bspwm_and_dunst.sh
@@ -3,12 +3,17 @@
 # Source the colors from wal
 source "${HOME}/.cache/wal/colors.sh"
 
-# Set the border colors
-bspc config normal_border_color   "$color1"
-bspc config active_border_color   "$color2"
-bspc config focused_border_color  "$color15"
-bspc config presel_feedback_color "$color1"
+# Set the border colors if bspwm is running
+if [ -n "$(wmctrl -m | grep 'bspwm')" ]; then
+  bspc config normal_border_color   "$color1"
+  bspc config active_border_color   "$color2"
+  bspc config focused_border_color  "$color15"
+  bspc config presel_feedback_color "$color1"
+fi
 
-# Restart dunst with the new color scheme
-pkill dunst
-dunst -conf "${HOME}/.cache/wal/dunstrc"
+# Restart dunst with the new color scheme, only if
+# we're not using Plasma
+if ! [ "$DESKTOP_SESSION" == "Plasma" ]; then
+  pkill dunst
+  dunst -conf "${HOME}/.cache/wal/dunstrc"
+fi
